// cliRepoPath/cms/src/views/page-management-versions/constants/page-management-enum.ts
export enum PageStatusEnum {
    ACTIVE = 'ACTIVE',
    INACTIVE = 'INACTIVE'
}

export enum PaginationSortEnum {
    ASC = 'asc',
    DESC = 'desc'
}

export enum AdminStatusEnum {
    ACTIVE = 'ACTIVE',
    VERIFIED = 'VERIFIED',
    PENDING = 'PENDING',
    INACTIVE = 'INACTIVE'
}

export enum AdminRolesTypeEnum {
    SUPER_ADMIN = 'SUPERADMIN',
    ADMIN = 'ADMIN'
}

export enum PageTypeEnum {
    ABOUT = 'ABOUT',
    GENERIC = 'GENERIC',
    CONTACT = 'CONTACT',
    TERMS_AND_CONDITION = 'TERMS_AND_CONDITION',
    PRIVACY_POLICY = 'PRIVACY_POLICY'
    // other titles
}